semiautonomous-agents/ms365-mcp-server/pyproject.toml
[project]
name = "ms365-mcp-server"
version = "0.1.0"
description = "Secure Microsoft 365 MCP Server for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.3.2",
    "msal>=1.28.0",
    "msgraph-sdk>=1.4.0",
    "google-cloud-secret-manager>=2.18.0",
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]
include = ["server.py", "auth.py", "graph_client.py"]
